requirements.txt
Pillow==10.2.0
numpy==1.26.4
qiskit==1.0.2
qiskit-aer==0.13.3 